Don't try and cleanup a directory that doesn't exist (in badging tests)

+++ lms/djangoapps/badges/tests/test_models.py
@@ -77,9 +77,10 @@ class BadgeClassTest(ModuleStoreTestCase):
         Remove all files uploaded as badges.
         """
         upload_to = BadgeClass._meta.get_field('image').upload_to  # pylint: disable=protected-access
-        (_, files) = default_storage.listdir(upload_to)
-        for uploaded_file in files:
-            default_storage.delete(upload_to + '/' + uploaded_file)
+        if default_storage.exists(upload_to):
+            (_, files) = default_storage.listdir(upload_to)
+            for uploaded_file in files:
+                default_storage.delete(upload_to + '/' + uploaded_file)
 
     # Need full path to make sure class names line up.
     @override_settings(BADGING_BACKEND='lms.djangoapps.badges.tests.test_models.DummyBackend')
